+++ src/args.rs
@@ -30,9 +30,9 @@ pub struct Args {
 pub enum Subcommands {
     /// Configure the source files to import the bookmarks.
     Config(ConfigArgs),
-    /// Determine diff of source and target bookmarks. Fetch and cache websites
-    /// for new bookmarks; delete cache for removed bookmarks.
-    Update(UpdateArgs),
+    /// Synchronize source and target bookmarks. Fetch and cache websites for
+    /// new bookmarks; delete cache for removed bookmarks.
+    Sync(SyncArgs),
     /// Import bookmarks from the configured source files.
     Import(ImportArgs),
     /// Fetch and cache bookmarks.
@@ -59,6 +59,16 @@ pub struct ConfigArgs {
     pub set_ignored_urls: SetIgnoredUrls,
     #[command(flatten)]
     pub set_underlying_urls: SetUnderlyingUrls,
+    #[command(flatten)]
+    pub set_request_timeout: SetRequestTimeout,
+    #[command(flatten)]
+    pub set_request_throttling: SetRequestThrottling,
+    #[command(flatten)]
+    pub set_max_concurrent_requests: SetMaxConcurrentRequests,
+    #[command(flatten)]
+    pub set_max_idle_connections_per_host: SetMaxIdleConnectionsPerHost,
+    #[command(flatten)]
+    pub set_idle_connections_timeout: SetIdleConnectionsTimeout,
 }
 
 #[derive(ClapArgs, Debug)]
@@ -96,6 +106,41 @@ pub struct SetUnderlyingUrls {
     pub underlying: Vec<String>,
 }
 
+#[derive(ClapArgs, Debug)]
+#[group(required = false)]
+pub struct SetRequestTimeout {
+    #[arg(long)]
+    pub request_timeout: Option<u64>,
+}
+
+#[derive(ClapArgs, Debug)]
+#[group(required = false)]
+pub struct SetRequestThrottling {
+    #[arg(long)]
+    pub request_throttling: Option<u64>,
+}
+
+#[derive(ClapArgs, Debug)]
+#[group(required = false)]
+pub struct SetMaxConcurrentRequests {
+    #[arg(long)]
+    pub max_concurrent_requests: Option<usize>,
+}
+
+#[derive(ClapArgs, Debug)]
+#[group(required = false)]
+pub struct SetMaxIdleConnectionsPerHost {
+    #[arg(long)]
+    pub max_idle_connections_per_host: Option<usize>,
+}
+
+#[derive(ClapArgs, Debug)]
+#[group(required = false)]
+pub struct SetIdleConnectionsTimeout {
+    #[arg(long)]
+    pub idle_connections_timeout: Option<u64>,
+}
+
 /// Describes the arguments for the `import` subcommand.
 #[derive(ClapArgs, Debug)]
 pub struct ImportArgs {
@@ -107,13 +152,12 @@ pub struct ImportArgs {
 /// Describes the arguments for the `fetch` subcommand.
 #[derive(ClapArgs, Debug)]
 pub struct FetchArgs {
-    /// Fetch all bookmarks.
+    /// Fetch and replace bookmarks.
     ///
-    /// If flag is not set, bookmarks are only fetched if a bookmark is not
-    /// cached yet. Otherwise, the cached content will be updated with
-    /// the fetched content.
+    /// If the flag is set, existing bookmarks will be fetched, and
+    /// the cached content will be replaced with the fetched content.
     #[arg(short, long)]
-    pub all: bool,
+    pub replace: bool,
     /// Cache the fetched bookmarks as text, HTML or markdown file.
     #[arg(short, long, value_enum)]
     pub mode: Option<CacheMode>,
@@ -134,9 +178,9 @@ pub struct FetchArgs {
     pub dry_run: bool,
 }
 
-/// Describes the arguments for the `update` subcommand.
+/// Describes the arguments for the `sync` subcommand.
 #[derive(ClapArgs, Debug)]
-pub struct UpdateArgs {
+pub struct SyncArgs {
     /// Cache the fetched bookmarks as text, HTML or markdown file.
     #[arg(short, long, value_enum)]
     pub mode: Option<CacheMode>,