+++ src/cmd/add.rs
@@ -9,6 +9,7 @@ use anyhow::anyhow;
 use chrono::Utc;
 use log::debug;
 
+/// Add urls to bookmarks.
 pub async fn add(config: Config, args: AddArgs) -> Result<(), anyhow::Error> {
     debug!("{args:?}");
 
@@ -26,19 +27,18 @@ pub async fn add(config: Config, args: AddArgs) -> Result<(), anyhow::Error> {
     )?;
     let client_config = ClientConfig::new(&config.settings);
     let cache_mode = CacheMode::new(&None, &config.settings.cache_mode);
+    let cache = Cache::new(&config.cache_path, cache_mode);
+    let client = Client::new(&client_config)?;
     let target_reader_writer = TargetReaderWriter::new(
         &config.target_bookmark_file,
         &config.target_bookmark_lock_file,
     )?;
-    let cache = Cache::new(&config.cache_path, cache_mode);
-    let client = Client::new(&client_config)?;
     let mut bookmark_manager = BookmarkManager::default();
     let bookmark_service = BookmarkService::new(service_config, client, cache);
 
     bookmark_service
         .run(
             &mut bookmark_manager,
-            &mut [],
             &mut target_reader_writer.reader(),
             &mut target_reader_writer.writer(),
             now,