+++ src/cmd/import.rs
@@ -1,6 +1,6 @@
 use crate::{
     args::ImportArgs,
-    bookmark_reader::{SourceReader, TargetReaderWriter},
+    bookmark_reader::TargetReaderWriter,
     bookmarks::{BookmarkManager, BookmarkService, RunMode, ServiceConfig},
     client::ClientConfig,
     cmd, json, utils, Cache, CacheMode, Client, Config,
@@ -38,17 +38,10 @@ pub async fn import(config: Config, args: ImportArgs) -> Result<(), anyhow::Erro
     let cache = Cache::new(&config.cache_path, cache_mode);
     let client_config = ClientConfig::new(&config.settings);
     let client = Client::new(&client_config)?;
-    let mut source_readers = config
-        .settings
-        .sources
-        .iter()
-        .map(SourceReader::init)
-        .collect::<Result<Vec<_>, anyhow::Error>>()?;
     let target_reader_writer = TargetReaderWriter::new(
         &config.target_bookmark_file,
         &config.target_bookmark_lock_file,
     )?;
-
     let now = Utc::now();
     let run_mode = if args.dry_run {
         RunMode::DryRun
@@ -60,13 +53,12 @@ pub async fn import(config: Config, args: ImportArgs) -> Result<(), anyhow::Erro
         &config.settings.ignored_urls,
         config.settings.max_concurrent_requests,
     )?;
-    let mut bookmark_manager = BookmarkManager::default();
+    let mut bookmark_manager = BookmarkManager::from_sources(&config.settings.sources)?;
     let bookmark_service = BookmarkService::new(service_config, client, cache);
 
     bookmark_service
         .run(
             &mut bookmark_manager,
-            &mut source_readers,
             &mut target_reader_writer.reader(),
             &mut target_reader_writer.writer(),
             now,