+++ src/cmd/remove.rs
@@ -9,6 +9,7 @@ use anyhow::anyhow;
 use chrono::Utc;
 use log::debug;
 
+/// Remove urls from bookmarks.
 pub async fn remove(config: Config, args: RemoveArgs) -> Result<(), anyhow::Error> {
     debug!("{args:?}");
 
@@ -19,10 +20,6 @@ pub async fn remove(config: Config, args: RemoveArgs) -> Result<(), anyhow::Erro
     }
 
     let now = Utc::now();
-    let target_reader_writer = TargetReaderWriter::new(
-        &config.target_bookmark_file,
-        &config.target_bookmark_lock_file,
-    )?;
     let service_config = ServiceConfig::new(
         RunMode::RemoveUrls(urls.clone()),
         &[],
@@ -32,13 +29,16 @@ pub async fn remove(config: Config, args: RemoveArgs) -> Result<(), anyhow::Erro
     let cache_mode = CacheMode::new(&None, &config.settings.cache_mode);
     let cache = Cache::new(&config.cache_path, cache_mode);
     let client = Client::new(&client_config)?;
+    let target_reader_writer = TargetReaderWriter::new(
+        &config.target_bookmark_file,
+        &config.target_bookmark_lock_file,
+    )?;
     let mut bookmark_manager = BookmarkManager::default();
     let bookmark_service = BookmarkService::new(service_config, client, cache);
 
     bookmark_service
         .run(
             &mut bookmark_manager,
-            &mut [],
             &mut target_reader_writer.reader(),
             &mut target_reader_writer.writer(),
             now,