+++ src/cmd/sync.rs
@@ -1,6 +1,6 @@
 use crate::{
-    args::UpdateArgs,
-    bookmark_reader::{SourceReader, TargetReaderWriter},
+    args::SyncArgs,
+    bookmark_reader::TargetReaderWriter,
     bookmarks::{BookmarkManager, BookmarkService, RunMode, ServiceConfig},
     cache::CacheMode,
     client::ClientConfig,
@@ -11,7 +11,7 @@ use log::debug;
 
 /// Import the diff of source and target bookmarks. Fetch and cache websites for
 /// new bookmarks; delete cache for removed bookmarks.
-pub async fn update(config: &Config, args: &UpdateArgs) -> Result<(), anyhow::Error> {
+pub async fn sync(config: &Config, args: &SyncArgs) -> Result<(), anyhow::Error> {
     debug!("{args:?}");
 
     if args.dry_run {
@@ -23,12 +23,6 @@ pub async fn update(config: &Config, args: &UpdateArgs) -> Result<(), anyhow::Er
     let client_config = ClientConfig::new(&config.settings);
     let client = Client::new(&client_config)?;
 
-    let mut source_readers = config
-        .settings
-        .sources
-        .iter()
-        .map(SourceReader::init)
-        .collect::<Result<Vec<_>, anyhow::Error>>()?;
     let target_reader_writer = TargetReaderWriter::new(
         &config.target_bookmark_file,
         &config.target_bookmark_lock_file,
@@ -37,20 +31,19 @@ pub async fn update(config: &Config, args: &UpdateArgs) -> Result<(), anyhow::Er
     let run_mode = if args.dry_run {
         RunMode::DryRun
     } else {
-        RunMode::Update
+        RunMode::Sync
     };
     let service_config = ServiceConfig::new(
         run_mode,
         &config.settings.ignored_urls,
         config.settings.max_concurrent_requests,
     )?;
-    let mut bookmark_manager = BookmarkManager::default();
+    let mut bookmark_manager = BookmarkManager::from_sources(&config.settings.sources)?;
     let bookmark_service = BookmarkService::new(service_config, client, cache);
 
     bookmark_service
         .run(
             &mut bookmark_manager,
-            &mut source_readers,
             &mut target_reader_writer.reader(),
             &mut target_reader_writer.writer(),
             now,