+++ tests/test_clean.rs
@@ -10,6 +10,7 @@ use tempfile::tempdir;
 
 #[tokio::test]
 async fn test_clean() {
+    let request_throttling = "1";
     let mock_server = common::start_mock_server().await;
     let mocks = common::mount_mocks(&mock_server, 3).await;
     let temp_dir = tempdir().unwrap();
@@ -24,29 +25,42 @@ async fn test_clean() {
         writeln!(file, "{}", url).unwrap();
     }
 
-    println!("Execute 'bogrep config --source {}'", source.display());
+    println!(
+        "Execute 'bogrep config --source {} --request-throttling {request_throttling}'",
+        source.display()
+    );
     let mut cmd = Command::cargo_bin(env!("CARGO_PKG_NAME")).unwrap();
     cmd.env("BOGREP_HOME", temp_path);
-    cmd.args(["config", "--source", source.to_str().unwrap()]);
-    cmd.output().unwrap();
+    cmd.args([
+        "config",
+        "--source",
+        source.to_str().unwrap(),
+        "--request-throttling",
+        request_throttling,
+    ]);
+    let res = cmd.output();
+    assert!(res.is_ok(), "Can't execute command: {}", res.unwrap_err());
 
     println!("Execute 'bogrep import'");
     let mut cmd = Command::cargo_bin(env!("CARGO_PKG_NAME")).unwrap();
     cmd.env("BOGREP_HOME", temp_path);
     cmd.args(["import"]);
-    cmd.output().unwrap();
+    let res = cmd.output();
+    assert!(res.is_ok(), "Can't execute command: {}", res.unwrap_err());
 
     println!("Execute 'bogrep fetch --mode text'");
     let mut cmd = Command::cargo_bin(env!("CARGO_PKG_NAME")).unwrap();
     cmd.env("BOGREP_HOME", temp_path);
     cmd.args(["fetch", "--mode", "text"]);
-    cmd.output().unwrap();
+    let res = cmd.output();
+    assert!(res.is_ok(), "Can't execute command: {}", res.unwrap_err());
 
     println!("Execute 'bogrep fetch --mode html'");
     let mut cmd = Command::cargo_bin(env!("CARGO_PKG_NAME")).unwrap();
     cmd.env("BOGREP_HOME", temp_path);
     cmd.args(["fetch", "--mode", "html"]);
-    cmd.output().unwrap();
+    let res = cmd.output();
+    assert!(res.is_ok(), "Can't execute command: {}", res.unwrap_err());
 
     let bookmarks_path = temp_dir.path().join("bookmarks.json");
     let bookmarks = utils::read_file(&bookmarks_path).unwrap();
@@ -79,6 +93,7 @@ async fn test_clean() {
 
 #[tokio::test]
 async fn test_clean_all() {
+    let request_throttling = "1";
     let mock_server = common::start_mock_server().await;
     let mocks = common::mount_mocks(&mock_server, 3).await;
     let temp_dir = tempdir().unwrap();
@@ -94,29 +109,42 @@ async fn test_clean_all() {
         writeln!(file, "{}", url).unwrap();
     }
 
-    println!("Execute 'bogrep config --source {}'", source.display());
+    println!(
+        "Execute 'bogrep config --source {} --request-throttling {request_throttling}'",
+        source.display()
+    );
     let mut cmd = Command::cargo_bin(env!("CARGO_PKG_NAME")).unwrap();
     cmd.env("BOGREP_HOME", temp_path);
-    cmd.args(["config", "--source", source.to_str().unwrap()]);
-    cmd.output().unwrap();
+    cmd.args([
+        "config",
+        "--source",
+        source.to_str().unwrap(),
+        "--request-throttling",
+        request_throttling,
+    ]);
+    let res = cmd.output();
+    assert!(res.is_ok(), "Can't execute command: {}", res.unwrap_err());
 
     println!("Execute 'bogrep import'");
     let mut cmd = Command::cargo_bin(env!("CARGO_PKG_NAME")).unwrap();
     cmd.env("BOGREP_HOME", temp_path);
     cmd.args(["import"]);
-    cmd.output().unwrap();
+    let res = cmd.output();
+    assert!(res.is_ok(), "Can't execute command: {}", res.unwrap_err());
 
     println!("Execute 'bogrep fetch --mode text'");
     let mut cmd = Command::cargo_bin(env!("CARGO_PKG_NAME")).unwrap();
     cmd.env("BOGREP_HOME", temp_path);
     cmd.args(["fetch", "--mode", "text"]);
-    cmd.output().unwrap();
+    let res = cmd.output();
+    assert!(res.is_ok(), "Can't execute command: {}", res.unwrap_err());
 
     println!("Execute 'bogrep fetch --mode html'");
     let mut cmd = Command::cargo_bin(env!("CARGO_PKG_NAME")).unwrap();
     cmd.env("BOGREP_HOME", temp_path);
     cmd.args(["fetch", "--mode", "html"]);
-    cmd.output().unwrap();
+    let res = cmd.output();
+    assert!(res.is_ok(), "Can't execute command: {}", res.unwrap_err());
 
     let bookmarks_path = temp_dir.path().join("bookmarks.json");
     let bookmarks = utils::read_file(&bookmarks_path).unwrap();