+++ tests/test_search.rs
@@ -10,6 +10,7 @@ use tempfile::tempdir;
 
 #[tokio::test]
 async fn test_search_case_sensitive() {
+    let request_throttling = "1";
     let mock_server = common::start_mock_server().await;
     let mocks = common::mount_mocks(&mock_server, 3).await;
     let temp_dir = tempdir().unwrap();
@@ -24,23 +25,35 @@ async fn test_search_case_sensitive() {
         writeln!(file, "{}", url).unwrap();
     }
 
-    println!("Execute 'bogrep config --source {}'", source.display());
+    println!(
+        "Execute 'bogrep config --source {} --request-throttling {request_throttling}'",
+        source_path.display()
+    );
     let mut cmd = Command::cargo_bin(env!("CARGO_PKG_NAME")).unwrap();
     cmd.env("BOGREP_HOME", temp_path);
-    cmd.args(["config", "--source", source.to_str().unwrap()]);
-    cmd.output().unwrap();
+    cmd.args([
+        "config",
+        "--source",
+        source.to_str().unwrap(),
+        "--request-throttling",
+        request_throttling,
+    ]);
+    let res = cmd.output();
+    assert!(res.is_ok(), "Can't execute command: {}", res.unwrap_err());
 
     println!("Execute 'bogrep import'");
     let mut cmd = Command::cargo_bin(env!("CARGO_PKG_NAME")).unwrap();
     cmd.env("BOGREP_HOME", temp_path);
     cmd.args(["import"]);
-    cmd.output().unwrap();
+    let res = cmd.output();
+    assert!(res.is_ok(), "Can't execute command: {}", res.unwrap_err());
 
     println!("Execute 'bogrep fetch'");
     let mut cmd = Command::cargo_bin(env!("CARGO_PKG_NAME")).unwrap();
     cmd.env("BOGREP_HOME", temp_path);
     cmd.args(["fetch"]);
-    cmd.output().unwrap();
+    let res = cmd.output();
+    assert!(res.is_ok(), "Can't execute command: {}", res.unwrap_err());
 
     println!("Execute 'bogrep \"Test content 1\"'");
     let mut cmd = Command::cargo_bin(env!("CARGO_PKG_NAME")).unwrap();
@@ -67,6 +80,7 @@ async fn test_search_case_sensitive() {
 
 #[tokio::test]
 async fn test_search_case_insensitive() {
+    let request_throttling = "1";
     let mock_server = common::start_mock_server().await;
     let mocks = common::mount_mocks(&mock_server, 3).await;
     let temp_dir = tempdir().unwrap();
@@ -81,23 +95,35 @@ async fn test_search_case_insensitive() {
         writeln!(file, "{}", url).unwrap();
     }
 
-    println!("Execute 'bogrep config --source {}'", source.display());
+    println!(
+        "Execute 'bogrep config --source {} --request-throttling {request_throttling}'",
+        source_path.display()
+    );
     let mut cmd = Command::cargo_bin(env!("CARGO_PKG_NAME")).unwrap();
     cmd.env("BOGREP_HOME", temp_path);
-    cmd.args(["config", "--source", source.to_str().unwrap()]);
-    cmd.output().unwrap();
+    cmd.args([
+        "config",
+        "--source",
+        source.to_str().unwrap(),
+        "--request-throttling",
+        request_throttling,
+    ]);
+    let res = cmd.output();
+    assert!(res.is_ok(), "Can't execute command: {}", res.unwrap_err());
 
     println!("Execute 'bogrep import'");
     let mut cmd = Command::cargo_bin(env!("CARGO_PKG_NAME")).unwrap();
     cmd.env("BOGREP_HOME", temp_path);
     cmd.args(["import"]);
-    cmd.output().unwrap();
+    let res = cmd.output();
+    assert!(res.is_ok(), "Can't execute command: {}", res.unwrap_err());
 
     println!("Execute 'bogrep fetch'");
     let mut cmd = Command::cargo_bin(env!("CARGO_PKG_NAME")).unwrap();
     cmd.env("BOGREP_HOME", temp_path);
     cmd.args(["fetch"]);
-    cmd.output().unwrap();
+    let res = cmd.output();
+    assert!(res.is_ok(), "Can't execute command: {}", res.unwrap_err());
 
     println!("Execute 'bogrep -i \"test content 1\"'");
     let mut cmd = Command::cargo_bin(env!("CARGO_PKG_NAME")).unwrap();
@@ -124,6 +150,7 @@ async fn test_search_case_insensitive() {
 
 #[tokio::test]
 async fn test_search_no_content() {
+    let request_throttling = "1";
     let mock_server = common::start_mock_server().await;
     let mocks = common::mount_mocks(&mock_server, 3).await;
     let temp_dir = tempdir().unwrap();
@@ -138,23 +165,35 @@ async fn test_search_no_content() {
         writeln!(file, "{}", url).unwrap();
     }
 
-    println!("Execute 'bogrep config --source {}'", source.display());
+    println!(
+        "Execute 'bogrep config --source {} --request-throttling {request_throttling}'",
+        source_path.display()
+    );
     let mut cmd = Command::cargo_bin(env!("CARGO_PKG_NAME")).unwrap();
     cmd.env("BOGREP_HOME", temp_path);
-    cmd.args(["config", "--source", source.to_str().unwrap()]);
-    cmd.output().unwrap();
+    cmd.args([
+        "config",
+        "--source",
+        source.to_str().unwrap(),
+        "--request-throttling",
+        request_throttling,
+    ]);
+    let res = cmd.output();
+    assert!(res.is_ok(), "Can't execute command: {}", res.unwrap_err());
 
     println!("Execute 'bogrep import'");
     let mut cmd = Command::cargo_bin(env!("CARGO_PKG_NAME")).unwrap();
     cmd.env("BOGREP_HOME", temp_path);
     cmd.args(["import"]);
-    cmd.output().unwrap();
+    let res = cmd.output();
+    assert!(res.is_ok(), "Can't execute command: {}", res.unwrap_err());
 
     println!("Execute 'bogrep fetch'");
     let mut cmd = Command::cargo_bin(env!("CARGO_PKG_NAME")).unwrap();
     cmd.env("BOGREP_HOME", temp_path);
     cmd.args(["fetch"]);
-    cmd.output().unwrap();
+    let res = cmd.output();
+    assert!(res.is_ok(), "Can't execute command: {}", res.unwrap_err());
 
     println!("Execute 'bogrep -l \"Test content 1\"'");
     let mut cmd = Command::cargo_bin(env!("CARGO_PKG_NAME")).unwrap();